Move more 3pt implemenations to Corr3 class
