+++ treecorr/kkgcorrelation.py
@@ -20,7 +20,6 @@ import numpy as np
 from . import _treecorr
 from .catalog import calculateVarK, calculateVarG
 from .corr3base import Corr3
-from .util import make_writer, make_reader
 from .config import make_minimal_config
 
 
@@ -81,9 +80,7 @@ class KKGCorrelation(Corr3):
     _sig3 = 'sig_sn (per component)'
 
     def __init__(self, config=None, *, logger=None, **kwargs):
-        """Initialize `KKGCorrelation`.  See class doc for details.
-        """
-        Corr3.__init__(self, config, logger=logger, **kwargs)
+        super().__init__(config, logger=logger, **kwargs)
 
         shape = self.data_shape
         self._z[0:2] = [np.zeros(shape, dtype=float) for _ in range(2)]
@@ -102,42 +99,13 @@ class KKGCorrelation(Corr3):
     def zetai(self):
         return self._z[1]
 
-    def process_cross21(self, cat1, cat2, *, metric=None, ordered=True, num_threads=None):
-        """Process two catalogs, accumulating the 3pt cross-correlation, where two of the
-        points in each triangle come from the first catalog, and one comes from the second.
-
-        This accumulates the cross-correlation for the given catalogs as part of a larger
-        auto- or cross-correlation calculation.  E.g. when splitting up a large catalog into
-        patches, this is appropriate to use for the cross correlation between different patches
-        as part of the complete auto-correlation of the full catalog.
-
-        Parameters:
-            cat1 (Catalog):     The first catalog to process. (2 points in each triangle will come
-                                from this catalog.)
-            cat2 (Catalog):     The second catalog to process. (1 point in each triangle will come
-                                from this catalog.)
-            metric (str):       Which metric to use.  See `Metrics` for details.
-                                (default: 'Euclidean'; this value can also be given in the
-                                constructor in the config dict.)
-            ordered (bool):     Whether to fix the order of the triangle vertices to match the
-                                catalogs. (default: True)
-            num_threads (int):  How many OpenMP threads to use during the calculation.
-                                (default: use the number of cpu cores; this value can also be given
-                                in the constructor in the config dict.)
-        """
-        super()._process_cross21(cat1, cat2, metric, ordered, num_threads)
-
     def finalize(self, vark1, vark2, varg):
         """Finalize the calculation of the correlation function.
 
-        The `process_auto`, `process_cross12` and `Corr3.process_cross` commands accumulate values
-        in each bin, so they can be called multiple times if appropriate.  Afterwards, this command
-        finishes the calculation by dividing by the total weight.
-
         Parameters:
-            vark1 (float):  The variance per component of the first scalar field.
-            vark2 (float):  The variance per component of the second scalar field.
-            varg (float):  The variance per component of the shear field.
+            vark1 (float):  The variance of the first scalar field.
+            vark2 (float):  The variance of the second scalar field.
+            varg (float):   The variance per component of the shear field.
         """
         self._finalize()
         mask1 = self.weightr != 0
@@ -153,256 +121,32 @@ class KKGCorrelation(Corr3):
     @property
     def varzeta(self):
         if self._varzeta is None:
-            self._varzeta = np.zeros(self.data_shape)
-            if self._var_num != 0:
-                self._varzeta.ravel()[:] = self.cov_diag[0:self._nbins].real
+            self._varzeta = self._calculate_varzeta(self._varzeta)
         return self._varzeta
 
     def _clear(self):
         super()._clear()
         self._varzeta = None
 
-    def _sum(self, others):
-        # Equivalent to the operation of:
-        #     self._clear()
-        #     for other in others:
-        #         self += other
-        # but no sanity checks and use numpy.sum for faster calculation.
-        np.sum([c.zetar for c in others], axis=0, out=self.zetar)
-        np.sum([c.zetai for c in others], axis=0, out=self.zetai)
-        np.sum([c.meand1 for c in others], axis=0, out=self.meand1)
-        np.sum([c.meanlogd1 for c in others], axis=0, out=self.meanlogd1)
-        np.sum([c.meand2 for c in others], axis=0, out=self.meand2)
-        np.sum([c.meanlogd2 for c in others], axis=0, out=self.meanlogd2)
-        np.sum([c.meand3 for c in others], axis=0, out=self.meand3)
-        np.sum([c.meanlogd3 for c in others], axis=0, out=self.meanlogd3)
-        np.sum([c.meanu for c in others], axis=0, out=self.meanu)
-        if self.bin_type == 'LogRUV':
-            np.sum([c.meanv for c in others], axis=0, out=self.meanv)
-        np.sum([c.weightr for c in others], axis=0, out=self.weightr)
-        if self.bin_type == 'LogMultipole':
-            np.sum([c.weighti for c in others], axis=0, out=self.weighti)
-        np.sum([c.ntri for c in others], axis=0, out=self.ntri)
-
-    def toSAS(self, *, target=None, **kwargs):
-        """Convert a multipole-binned correlation to the corresponding SAS binning.
-
-        This is only valid for bin_type == LogMultipole.
-
-        Keyword Arguments:
-            target:     A target KKGCorrelation object with LogSAS binning to write to.
-                        If this is not given, a new object will be created based on the
-                        configuration paramters of the current object. (default: None)
-            **kwargs:   Any kwargs that you want to use to configure the returned object.
-                        Typically, might include min_phi, max_phi, nphi_bins, phi_bin_size.
-                        The default phi binning is [0,pi] with nphi_bins = self.max_n.
-
-        Returns:
-            A KKGCorrelation object with bin_type=LogSAS containing the
-            same information as this object, but with the SAS binning.
-        """
-        sas = super().toSAS(target=target, **kwargs)
-
-        sas._var_num = self._var_num
-
-        # Z(d2,d3,phi) = 1/2pi sum_n Z_n(d2,d3) exp(i n phi)
-        expiphi = np.exp(1j * self.n1d[:,None] * sas.phi1d)
-        zeta = self.zeta.dot(expiphi) / (2*np.pi) * sas.phi_bin_size
-
-        # We leave the gam_mu unnormalized in the Multipole class, so after the FT,
-        # we still need to divide by weight.
-        mask = sas.weightr != 0
-        zeta[mask] /= sas.weightr[mask]
-
-        # Now fix the projection.
-        # The multipole algorithm uses the Porth et al x projection.
-        # We need to switch that to the canoical centroid projection.
-
-        # Define some complex "vectors" where p1 is at the origin and
-        # p3 is on the x axis:
-        # s = p3 - p1
-        # t = p2 - p1
-        # u = angle bisector of s, t
-        # q1 = (s+t)/3.  (this is the centroid)
-        # q2 = q1-t
-        # q3 = q1-s
-        s = sas.meand2
-        t = sas.meand3 * np.exp(1j * sas.meanphi * sas._phi_units)
-        u = (1 + t/np.abs(t))/2
-        q1 = (s+t)/3.
-        q2 = q1-t
-        q3 = q1-s
-
-        # Currently the projection is as follows:
-        # g1 is projected along u
-        # g2 is projected along t
-        # g3 is projected along s
-        #
-        # We want to have
-        # g1 projected along q1
-        # g2 projected along q2
-        # g3 projected along q3
-        #
-        # The phases to multiply by are exp(2iphi_current) * exp(-2iphi_target). I.e.
-        # g1phase = (u conj(q1))**2 / |u conj(q1)|**2
-        # g2phase = (t conj(q2))**2 / |t conj(q2)|**2
-        # g3phase = (s conj(q3))**2 / |s conj(q3)|**2
-        g1phase = (u * np.conj(q1))**2
-        g2phase = (t * np.conj(q2))**2
-        g3phase = (s * np.conj(q3))**2
-        g1phase /= np.abs(g1phase)
-        g2phase /= np.abs(g2phase)
-        g3phase /= np.abs(g3phase)
-
-        # Now just multiply each gam by the appropriate combination of phases.
-        # FIXME
-        zetaphase = g1phase * g2phase * g3phase
-        zeta *= zetaphase
-
-        sas.zetar[:] = np.real(zeta)
-        sas.zetai[:] = np.imag(zeta)
-
-        for k,v in self.results.items():
-            temp = sas.results[k]
-            zeta = v.zeta.dot(expiphi) / (2*np.pi) * sas.phi_bin_size * zetaphase
-            temp.zetar[:] = np.real(zeta)
-            temp.zetai[:] = np.imag(zeta)
-
-        return sas
-
     def write(self, file_name, *, file_type=None, precision=None, write_patch_results=False,
               write_cov=False):
-        r"""Write the correlation function to the file, file_name.
+        super().write(file_name, file_type=file_type, precision=precision,
+                      write_patch_results=write_patch_results, write_cov=write_cov)
 
-        As described in the doc string for `KKGCorrelation`, we use the "natural components" of
-        the shear 3-point function described by Schneider & Lombardi (2003) using the triangle
-        centroid as the projection point.  There are 4 complex-valued natural components, so there
-        are 8 columns in the output file.
-
-        For bin_type = LogRUV, the output file will include the following columns:
-
-        ==========      ================================================================
-        Column          Description
-        ==========      ================================================================
-        r_nom           The nominal center of the bin in r = d2 where d1 > d2 > d3
-        u_nom           The nominal center of the bin in u = d3/d2
-        v_nom           The nominal center of the bin in v = +-(d1-d2)/d3
-        meanu           The mean value :math:`\langle u\rangle` of triangles that fell
-                        into each bin
-        meanv           The mean value :math:`\langle v\rangle` of triangles that fell
-                        into each bin
-        ==========      ================================================================
-
-        For bin_type = LogSAS, the output file will include the following columns:
-
-        ==========      ================================================================
-        Column          Description
-        ==========      ================================================================
-        d2_nom          The nominal center of the bin in d2
-        d3_nom          The nominal center of the bin in d3
-        phi_nom         The nominal center of the bin in phi, the opening angle between
-                        d2 and d3 in the counter-clockwise direction
-        meanphi         The mean value :math:`\langle phi\rangle` of triangles that fell
-                        into each bin
-        ==========      ================================================================
-
-        For bin_type = LogMultipole, the output file will include the following columns:
-
-        ==========      ================================================================
-        Column          Description
-        ==========      ================================================================
-        d2_nom          The nominal center of the bin in d2
-        d3_nom          The nominal center of the bin in d3
-        n               The multipole index n
-        ==========      ================================================================
-
-        In addition, all bin types include the following columns:
-
-        ==========      ================================================================
-        Column          Description
-        ==========      ================================================================
-        meand1          The mean value :math:`\langle d1\rangle` of triangles that fell
-                        into each bin
-        meanlogd1       The mean value :math:`\langle \log(d1)\rangle` of triangles that
-                        fell into each bin
-        meand2          The mean value :math:`\langle d2\rangle` of triangles that fell
-                        into each bin
-        meanlogd2       The mean value :math:`\langle \log(d2)\rangle` of triangles that
-                        fell into each bin
-        meand3          The mean value :math:`\langle d3\rangle` of triangles that fell
-                        into each bin
-        meanlogd3       The mean value :math:`\langle \log(d3)\rangle` of triangles that
-                        fell into each bin
+    write.__doc__ = Corr3.write.__doc__.format(
+        r"""
         zetar           The real part of the estimator of :math:`\Gamma_0`
         zetai           The imag part of the estimator of :math:`\Gamma_0`
         sigma_zeta      The sqrt of the variance estimate of :math:`\Gamma_0`
-        weight          The total weight of triangles contributing to each bin.
-                        (For LogMultipole, this is split into real and imaginary parts,
-                        weight_re and weight_im.)
-        ntri            The number of triangles contributing to each bin
-        ==========      ================================================================
-
-        If ``sep_units`` was given at construction, then the distances will all be in these units.
-        Otherwise, they will be in either the same units as x,y,z (for flat or 3d coordinates) or
-        radians (for spherical coordinates).
-
-        Parameters:
-            file_name (str):    The name of the file to write to.
-            file_type (str):    The type of file to write ('ASCII' or 'FITS').  (default: determine
-                                the type automatically from the extension of file_name.)
-            precision (int):    For ASCII output catalogs, the desired precision. (default: 4;
-                                this value can also be given in the constructor in the config dict.)
-            write_patch_results (bool): Whether to write the patch-based results as well.
-                                        (default: False)
-            write_cov (bool):   Whether to write the covariance matrix as well. (default: False)
-        """
-        self.logger.info('Writing KKG correlations to %s',file_name)
-        precision = self.config.get('precision', 4) if precision is None else precision
-        with make_writer(file_name, precision, file_type, self.logger) as writer:
-            self._write(writer, None, write_patch_results, write_cov=write_cov)
+        """)
 
     @property
-    def _write_col_names(self):
-        if self.bin_type == 'LogRUV':
-            col_names = ['r_nom', 'u_nom', 'v_nom',
-                         'meand1', 'meanlogd1', 'meand2', 'meanlogd2',
-                         'meand3', 'meanlogd3', 'meanu', 'meanv']
-        elif self.bin_type == 'LogSAS':
-            col_names = ['d2_nom', 'd3_nom', 'phi_nom',
-                         'meand1', 'meanlogd1', 'meand2', 'meanlogd2',
-                         'meand3', 'meanlogd3', 'meanphi']
-        else:
-            col_names = ['d2_nom', 'd3_nom', 'n',
-                         'meand1', 'meanlogd1', 'meand2', 'meanlogd2',
-                         'meand3', 'meanlogd3']
-        col_names += ['zetar', 'zetai', 'sigma_zeta']
-        if self.bin_type == 'LogMultipole':
-            col_names += ['weight_re', 'weight_im', 'ntri']
-        else:
-            col_names += ['weight', 'ntri']
-        return col_names
+    def _write_class_col_names(self):
+        return ['zetar', 'zetai', 'sigma_zeta']
 
     @property
-    def _write_data(self):
-        if self.bin_type == 'LogRUV':
-            data = [ self.rnom, self.u, self.v,
-                     self.meand1, self.meanlogd1, self.meand2, self.meanlogd2,
-                     self.meand3, self.meanlogd3, self.meanu, self.meanv ]
-        elif self.bin_type == 'LogSAS':
-            data = [ self.d2nom, self.d3nom, self.phi,
-                     self.meand1, self.meanlogd1, self.meand2, self.meanlogd2,
-                     self.meand3, self.meanlogd3, self.meanphi ]
-        else:
-            data = [ self.d2nom, self.d3nom, self.n,
-                     self.meand1, self.meanlogd1, self.meand2, self.meanlogd2,
-                     self.meand3, self.meanlogd3 ]
-        data += [ self.zetar, self.zetai, np.sqrt(self.varzeta) ]
-        if self.bin_type == 'LogMultipole':
-            data += [ self.weightr, self.weighti, self.ntri ]
-        else:
-            data += [ self.weight, self.ntri ]
-        data = [ col.flatten() for col in data ]
-        return data
+    def _write_class_data(self):
+        return [ self.zetar, self.zetai, np.sqrt(self.varzeta) ]
 
     def _read_from_data(self, data, params):
         super()._read_from_data(data, params)
