+++ treecorr/kvcorrelation.py
@@ -100,11 +100,6 @@ class KVCorrelation(BaseKZCorrelation):
     _xireal = 'xi'
     _xiimag = 'xi_im'
 
-    def __init__(self, config=None, *, logger=None, **kwargs):
-        """Initialize `KVCorrelation`.  See class doc for details.
-        """
-        super().__init__(config, logger=logger, **kwargs)
-
     def finalize(self, vark, varv):
         """Finalize the calculation of the correlation function.
