+++ treecorr/kzcorrelation.py
@@ -189,11 +189,6 @@ class KZCorrelation(BaseKZCorrelation):
     _xireal = 'xi'
     _xiimag = 'xi_im'
 
-    def __init__(self, config=None, *, logger=None, **kwargs):
-        """Initialize `KZCorrelation`.  See class doc for details.
-        """
-        super().__init__(config, logger=logger, **kwargs)
-
     def finalize(self, vark, varz):
         """Finalize the calculation of the correlation function.
