+++ treecorr/nkcorrelation.py
@@ -112,8 +112,6 @@ class NKCorrelation(Corr2):
     _default_angle_slop = 1
 
     def __init__(self, config=None, *, logger=None, **kwargs):
-        """Initialize `NKCorrelation`.  See class doc for details.
-        """
         super().__init__(config, logger=logger, **kwargs)
 
         self._xi1 = np.zeros_like(self.rnom, dtype=float)
